dsl: Add centre stencil functionality
