+++ devito/petsc/solve.py
@@ -6,8 +6,9 @@ from devito.finite_differences.differentiable import Add, Mul, EvalDerivative
 from devito.finite_differences.derivative import Derivative
 from devito.types import Eq
 from devito.operations.solve import eval_time_derivatives
+from devito.symbolics import retrieve_functions
+from devito.symbolics import uxreplace
 from devito.petsc.types import PETScArray, LinearSolveExpr, MatVecEq, RHSEq
-
 
 __all__ = ['PETScSolve']
 
@@ -26,14 +27,22 @@ def PETScSolve(eq, target, bcs=None, solver_parameters=None, **kwargs):
                    liveness='eager',
                    halo=target.halo[1:] if is_time_dep else target.halo)
         for prefix in ['y_matvec', 'x_matvec', 'b_tmp']]
+    
+    b, F_target = separate_eqn(eq, target)
+
+    # Args were updated so need to update target
+    new_target = {func for func in retrieve_functions(F_target) if 
+                  func.function == target.function}.pop()
+
+    centre = centre_stencil(Eq(y_matvec, F_target), new_target)
 
     # TODO: Extend to rearrange equation for implicit time stepping.
-    matvecaction = MatVecEq(y_matvec, LinearSolveExpr(eq.lhs.subs(target, x_matvec),
+    matvecaction = MatVecEq(y_matvec, LinearSolveExpr(uxreplace(F_target, {new_target: x_matvec}),
                             target=target, solver_parameters=solver_parameters),
                             subdomain=eq.subdomain)
 
     # Part of pde that remains constant at each timestep
-    rhs = RHSEq(b_tmp, LinearSolveExpr(eq.rhs, target=target,
+    rhs = RHSEq(b_tmp, LinearSolveExpr(b, target=target,
                 solver_parameters=solver_parameters), subdomain=eq.subdomain)
 
     if not bcs:
@@ -66,8 +75,9 @@ def separate_eqn(eqn, target):
     zeroed_eqn = Eq(eqn.lhs - eqn.rhs, 0)
     tmp = eval_time_derivatives(zeroed_eqn.lhs)
     b = remove_target(tmp, target)
-    F_target = simplify(tmp - b)
-
+    # Is this ok? Is there another way of using simplify
+    # but maintaining its Devito type?
+    F_target = Add(simplify(tmp - b))
     return -b, F_target
 
 
@@ -105,3 +115,42 @@ def _(expr, target):
 @remove_target.register(Derivative)
 def _(expr, target):
     return 0 if expr.has(target) else expr
+
+
+def centre_stencil(eqn, target):
+    """
+    Extract the centre stencil from equation.
+    """
+    new = extract_centre(eqn.evaluate.rhs, target)
+    return new
+
+
+@singledispatch
+def extract_centre(expr, target):
+    return expr if expr == target else 0
+
+
+@extract_centre.register(Add)
+@extract_centre.register(EvalDerivative)
+def _(expr, target):
+    if not expr.has(target):
+        return 0
+
+    args = [extract_centre(a, target) for a in expr.args]
+    return expr.func(*args, evaluate=False)
+
+
+@extract_centre.register(Mul)
+def _(expr, target):
+    if not expr.has(target):
+        return 0
+
+    args = []
+    for a in expr.args:
+        if not a.has(target):
+            args.append(a)
+        else:
+            a1 = extract_centre(a, target)
+            args.append(a1)
+
+    return expr.func(*args, evaluate=False)
